cebattle: Not % base

+++ CEBattle/CEBattle/WarMath.cs
@@ -16,8 +16,8 @@ namespace CEBattle
         /// </summary>
         /// <param name="side1">The number of strenght of side1</param>
         /// <param name="side2">The number of strenght of side2</param>
-        /// <returns>Return a following % of loss.  if value is less than 0, side 1 win, otherwise, side 2 win.</returns>
-        public static float ResultBalance(int side1, int side2)
+        /// <returns>Return a following impact of loss.  if value is less than 0, side 1 win, otherwise, side 2 win.</returns>
+        public static int ResultBalance(int side1, int side2)
         {
             // Step 1, use
             /*
@@ -30,11 +30,10 @@ namespace CEBattle
             // Random caught: 350
             // Means than side2 win.
             // 400 (300+100)-350 = 50
-            // 50/100 = 0.5
-            // Side 2 so 0.5 and not -0.5
+            // Side 2 so 50 and not -50
             // It means that the first skirmish was not that tight and the loss on side1 is pretty significative, at least 50 units on side one will be lost (Apply bonus later)
 
-            return 0.0f; // Just to compile
+            return 0; // Just to compile
         }
     }
 }
